fix(number-finder): replace with the working deployed version
The copy on main could not run. Three defects, all fixed by taking the version
from the local heroku repo, which is where this app was actually deployed from:

- render_template('index.html') resolves against templates/, but index.html sat
  at the package root — the app raised TemplateNotFound on every request.
- Procfile said 'web: python app.py' while the file was named number-finder.py.
- requirements.txt pinned wheel to an absolute
  file:///opt/homebrew/Cellar/... path, so 'pip install -r' failed on any
  machine but the one that wrote it.

The heroku version is also further developed:

- Precision is a form field (generate_pi(num_digits)) rather than mp.dps
  hardcoded at import. The old version computed a million digits of pi at module
  import, on every boot, whether or not anyone searched.
- Adds an upper bound (1,000,000,001) and thousands separators in the output.
- Adds a privacy disclaimer telling people not to search for sensitive data —
  worth having on a box that takes digit strings from strangers.
- Binds 0.0.0.0 on $PORT for Heroku instead of app.run(debug=True), which
  should never have been the production entrypoint.

Provenance: ~/Documents/dev/heroku, a git repo with 15 commits and no remote,
so this content existed nowhere else. Folding it here retires that repo.
READMEs were already identical.

Verified: app.py compiles, templates/index.html resolves, Procfile targets an
existing file, and requirements.txt has no local paths.
