+++ number-finder/requirements.txt
@@ -8,4 +8,3 @@ mclass==1.3.4
 mpath==1.1.3
 mpmath==1.3.0
 Werkzeug==3.0.4
-wheel @ file:///opt/homebrew/Cellar/python%403.12/3.12.6/libexec/wheel-0.44.0-py3-none-any.whl#sha256=d6ad895f80a568fbc066fd6c52c68c5ef40632c4d764d42af13e28e945f1920b
