+++ number-finder/templates/index.html
@@ -1,0 +1,103 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Find Sequence in Pi</title>
+    <!-- Add Bootstrap CSS for styling -->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet">
+    <style>
+        body {
+            background-color: #f8f9fa;
+            margin-top: 50px;
+        }
+        .container {
+            max-width: 600px;
+            background-color: #ffffff;
+            padding: 20px;
+            border-radius: 8px;
+            box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+        }
+        h1 {
+            color: #007bff;
+            text-align: center;
+        }
+        .form-group {
+            margin-bottom: 20px;
+        }
+        .disclaimer {
+            margin-top: 20px;
+            font-size: 0.9em;
+            color: #6c757d;
+        }
+        .contact-info {
+            margin-top: 40px;
+            font-size: 0.9em;
+            color: #343a40;
+            text-align: center;
+        }
+        .contact-info a {
+            color: #007bff;
+            text-decoration: none;
+        }
+        .contact-info a:hover {
+            text-decoration: underline;
+        }
+        /* Spinner styling */
+        .spinner-container {
+            display: none;
+            text-align: center;
+            margin-top: 20px;
+        }
+    </style>
+</head>
+<body>
+    <div class="container">
+        <h1>Find Sequence in Pi</h1>
+        <form method="POST" onsubmit="showSpinner()">
+            <div class="form-group">
+                <label for="number_str">Enter a string of digits (up to 9 digits):</label>
+                <input type="text" class="form-control" id="number_str" name="number_str" maxlength="9" required>
+            </div>
+            <div class="form-group">
+                <label for="num_digits">How many digits of Pi do you want to search? (up to 1,000,000,001):</label>
+                <input type="number" class="form-control" id="num_digits" name="num_digits" value="{{ num_digits }}" min="1" max="1000000001">
+            </div>
+            <button type="submit" class="btn btn-primary btn-block">Search in Pi</button>
+        </form>
+
+        <!-- Spinner that shows while searching -->
+        <div class="spinner-container" id="spinner">
+            <div class="spinner-border text-primary" role="status">
+                <span class="sr-only">Searching...</span>
+            </div>
+            <p>Searching, please wait...</p>
+        </div>
+
+        {% if result %}
+        <div class="mt-4">
+            <h2>Result:</h2>
+            <p>{{ result }}</p>
+        </div>
+        {% endif %}
+
+        <div class="disclaimer">
+            <p>{{ disclaimer }}</p>
+        </div>
+
+        <!-- Contact information at the bottom -->
+        <div class="contact-info">
+            <p>Curious to learn more? Contact <a href="mailto:email@dhk.io">email@dhk.io</a></p>
+        </div>
+    </div>
+
+    <!-- Bootstrap JS and custom script to show the spinner -->
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/js/bootstrap.bundle.min.js"></script>
+    <script>
+        // Function to show the spinner when the form is submitted
+        function showSpinner() {
+            document.getElementById('spinner').style.display = 'block';
+        }
+    </script>
+</body>
+</html>
